requirements.txt
fastapi
uvicorn
openai
tiktoken
pydantic
python-dotenv
langchain
chromadb
pypdf
sentence-transformers
loguru